+++ src/control/app/src/main/kotlin/org/calypsonet/keyple/demo/control/ui/ReaderActivity.kt
@@ -23,6 +23,7 @@ import kotlinx.coroutines.launch
 import kotlinx.coroutines.withContext
 import org.calypsonet.keyple.demo.control.R
 import org.calypsonet.keyple.demo.control.data.model.AppSettings
+import org.calypsonet.keyple.demo.control.data.model.AuthenticationMode
 import org.calypsonet.keyple.demo.control.data.model.CardReaderResponse
 import org.calypsonet.keyple.demo.control.data.model.Status
 import org.calypsonet.keyple.demo.control.databinding.ActivityCardReaderBinding
@@ -75,8 +76,8 @@ class ReaderActivity : BaseActivity() {
             ticketingService.init(cardReaderObserver, this@ReaderActivity, AppSettings.readerType)
             showToast(
                 getString(
-                    if (ticketingService.isSecureSessionMode) R.string.secure_session_mode_enabled
-                    else R.string.secure_session_mode_disabled))
+                    if (ticketingService.isSamAvailable) R.string.sam_available
+                    else R.string.sam_not_available))
             handleAppEvents(AppState.WAIT_CARD, null)
             ticketingService.startNfcDetection()
           } catch (e: Exception) {
@@ -134,7 +135,10 @@ class ReaderActivity : BaseActivity() {
           Timber.e("Card not selected: %s", error)
           displayResult(
               CardReaderResponse(
-                  status = Status.INVALID_CARD, titlesList = arrayListOf(), errorMessage = error))
+                  status = Status.INVALID_CARD,
+                  authenticationMode = AuthenticationMode.NO_AUTHENTICATION,
+                  titlesList = arrayListOf(),
+                  errorMessage = error))
           return
         }
         Timber.i("A Calypso Card selection succeeded.")
@@ -170,6 +174,14 @@ class ReaderActivity : BaseActivity() {
                       cardReaderResponse.status == Status.ERROR) {
                     ticketingService.displayResultFailed()
                   } else {
+                    when (cardReaderResponse.authenticationMode) {
+                      AuthenticationMode.SAM ->
+                          showToast(getString(R.string.authentication_mode_sam))
+                      AuthenticationMode.PKI ->
+                          showToast(getString(R.string.authentication_mode_pki))
+                      AuthenticationMode.NO_AUTHENTICATION ->
+                          showToast(getString(R.string.authentication_mode_no_authentication))
+                    }
                     ticketingService.displayResultSuccess()
                   }
                   progress.dismiss()
@@ -178,7 +190,11 @@ class ReaderActivity : BaseActivity() {
               } catch (e: IllegalStateException) {
                 Timber.e(e)
                 Timber.e("Load ERROR page after exception = ${e.message}")
-                displayResult(CardReaderResponse(status = Status.ERROR, titlesList = arrayListOf()))
+                displayResult(
+                    CardReaderResponse(
+                        status = Status.ERROR,
+                        authenticationMode = AuthenticationMode.NO_AUTHENTICATION,
+                        titlesList = arrayListOf()))
               }
             }
           }